test: consolidate duplicate root suites into nested counterparts (Closes #254)
Audit root-level duplicates against nested suites and merge only unique
coverage: ActionExecutor call tracking/limits, TaskRunner tool-error wrapping,
EventBus registration guards, and memory snapshot-stability. Delete redundant
root files whose assertions were already covered nested (logger levels,
RuntimeError toJSON, overlapping memory/property cases).
