+++ tests/events/runtime-events.test.ts
@@ -76,3 +76,38 @@ describe("RuntimeEventBus", () => {
     expect(otherListener).not.toHaveBeenCalled();
   });
 });
+
+describe("RuntimeEventBus registration guards (merged from root suite)", () => {
+  it("does not re-add the same listener when it is already registered", () => {
+    const eventBus = new RuntimeEventBus({ maxListeners: 1 });
+    const listener = vi.fn();
+
+    eventBus.on("runtime.started", listener);
+    expect(() => eventBus.on("runtime.started", listener)).not.toThrow();
+    expect(eventBus.listenerCount("runtime.started")).toBe(1);
+  });
+
+  it("applies maxListeners independently per event name", () => {
+    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
+    const eventBus = new RuntimeEventBus({ maxListeners: 1 });
+    eventBus.on("runtime.started", vi.fn());
+
+    // Second distinct listener for the same event warns but still registers.
+    expect(() => eventBus.on("runtime.started", vi.fn())).not.toThrow();
+    expect(warnSpy).toHaveBeenCalledTimes(1);
+    expect(eventBus.listenerCount("runtime.started")).toBe(2);
+
+    // A different event name has its own quota.
+    expect(() => eventBus.on("runtime.task.failed", vi.fn())).not.toThrow();
+    expect(eventBus.listenerCount("runtime.task.failed")).toBe(1);
+
+    warnSpy.mockRestore();
+  });
+
+  it.each([0, -1, 1.5, Number.NaN])(
+    "rejects invalid maxListeners value %s",
+    (maxListeners) => {
+      expect(() => new RuntimeEventBus({ maxListeners })).toThrow(RangeError);
+    }
+  );
+});
