+++ tests/memory/in-memory-memory-store.property.test.ts
@@ -10,7 +10,10 @@ describe("InMemoryMemoryStore property-based tests", () => {
   const taskIdArb = fc.string({ minLength: 1, maxLength: 20 });
   const inputArb = fc.string();
   const outputArb = fc.anything();
-  const recordedAtArb = fc.date().map((d) => d.toISOString());
+  const recordedAtArb = fc
+    .date({ noInvalidDate: true })
+    .filter((d) => !Number.isNaN(d.getTime()))
+    .map((d) => d.toISOString());
 
   const entryArb: fc.Arbitrary<MemoryEntry> = fc.record({
     agentId: agentIdArb,
@@ -86,6 +89,57 @@ describe("InMemoryMemoryStore property-based tests", () => {
             const result = await store.listByAgent(unknownAgent);
             expect(result).toEqual([]);
           }
+        }
+      ),
+      { numRuns: 50 }
+    );
+  });
+
+  it("later appends extend earlier snapshots without reordering them", async () => {
+    const AGENT_IDS = [
+      "agent-alpha",
+      "agent-beta",
+      "agent-gamma",
+      "agent-delta"
+    ] as const;
+    const boundedEntryArb: fc.Arbitrary<MemoryEntry> = fc.record({
+      agentId: fc.constantFrom(...AGENT_IDS),
+      taskId: fc.string({ maxLength: 32 }),
+      input: fc.string({ maxLength: 64 }),
+      output: fc.oneof(
+        fc.string({ maxLength: 64 }),
+        fc.integer(),
+        fc.boolean()
+      ),
+      recordedAt: recordedAtArb
+    });
+
+    await fc.assert(
+      fc.asyncProperty(
+        fc.array(boundedEntryArb, { maxLength: 120 }),
+        fc.constantFrom(...AGENT_IDS),
+        async (entries, probeAgent) => {
+          const store = new InMemoryMemoryStore();
+          const splitAt = Math.floor(entries.length / 2);
+
+          for (const entry of entries.slice(0, splitAt)) {
+            await store.append(entry);
+          }
+          const snapshot = await store.listByAgent(probeAgent);
+          expect(snapshot).toEqual(
+            entries
+              .slice(0, splitAt)
+              .filter((entry) => entry.agentId === probeAgent)
+          );
+
+          for (const entry of entries.slice(splitAt)) {
+            await store.append(entry);
+          }
+          const later = await store.listByAgent(probeAgent);
+          expect(later.slice(0, snapshot.length)).toEqual(snapshot);
+          expect(later).toEqual(
+            entries.filter((entry) => entry.agentId === probeAgent)
+          );
         }
       ),
       { numRuns: 50 }
